requirements.txt
streamlit>=1.33.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
torch>=2.1.0
torchvision>=0.16.0
imageio>=2.31.0
tqdm>=4.66.0
tensorflow